Reorganized java files into relevant packages, started groundwork for TaskList connecting to SharedPrefs
