+++ app/src/main/java/alexandertech/mymonteuniversityhub/Fragments/ParkingLocator.java
@@ -1,4 +1,4 @@
-package alexandertech.mymonteuniversityhub;
+package alexandertech.mymonteuniversityhub.Fragments;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,6 +6,8 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import alexandertech.mymonteuniversityhub.R;
 
 
 public class ParkingLocator extends Fragment {
