+++ app/src/main/java/alexandertech/mymonteuniversityhub/Methods/AddTask.java
@@ -1,0 +1,8 @@
+package alexandertech.mymonteuniversityhub.Methods;
+
+/**
+ * Created by aymswick on 3/10/17.
+ */
+
+public class AddTask {
+}
