+++ app/src/main/res/layout/fragment_planner.xml
@@ -15,12 +15,20 @@
             android:id="@+id/calendarView" />
     </android.support.v7.widget.CardView>
 
-    
+    <android.support.design.widget.FloatingActionButton
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/fab"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentRight="true"/>
+
+
     <android.support.v7.widget.RecyclerView
         android:id="@+id/cardList"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_below="@+id/calCard">
+        android:layout_below="@+id/calCard"
+        app:elevation="10dp">
 
     </android.support.v7.widget.RecyclerView>
